trailingNobitsSections: remove exception case where no nobits sections are found
This change removes an exception that was getting raised on a particular
input binary built for bare metal PowerPC. As far as I can tell, it
isn't clear to me that this error could be resolved in a meaningful way,
and removing it allows us to successfully process binaries that would
have triggered this error. It strikes me that it's a situation the
original author(s) thought might be pathological, but it isn't evident
to me that it should justify raising an exception.